refactor(storage): extract FileSystem trait+types into proximadb-storage-filesystem-types
Root-crate decomposition enabler (decomposition plan / TD-104). The FileSystem
and FilesystemFile traits plus their value types were a leaf seam buried in the
2073-line src/storage/persistence/filesystem/mod.rs. Several storage modules
(encryption, schema, common, metadata) depend on them, which forced those
modules to depend UP on the root crate — encryption<->filesystem was an actual
cycle — blocking their extraction as clean leaves.

Move the pure trait + value types (mod.rs lines 217-780) into a new leaf crate
crates/storage/proximadb-storage-filesystem-types and re-export them from the
old path (`pub use proximadb_storage_filesystem_types::*;`), so every existing
`crate::storage::persistence::filesystem::{FileSystem, ...}` consumer compiles
unchanged. The two trait impls orphaned by the move
(Default for FilesystemPerformanceConfig; From<FilesystemError> for
VectorDBError) move into the new crate where their types are local. The
engine/encryption-bound glue (FilesystemFactory, FilesystemConfig, the impl
submodules local/aws_s3/azure_blob/gcs_store/...) stays in root.

Behavior-neutral mechanical split (no logic change). Shrinks the root crate's
filesystem/mod.rs from 2073 to 1491 lines and gives encryption/schema/common/
metadata a downward dependency to extract against next.

Verified: cargo check --workspace --lib --bins (0 warnings); clippy --lib --bins
-D warnings; cargo fmt --check; scripts/check-layering.sh (no boundary findings,
new crate sits in the storage tier); server binaries build.
