+++ crates/storage/proximadb-storage-filesystem-types/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "proximadb-storage-filesystem-types"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage = "https://github.com/vjsingh1984/proximadb"
+description = "Pure filesystem abstraction (FileSystem/FilesystemFile traits + value types) for ProximaDB storage"
+
+[dependencies]
+async-trait.workspace = true
+chrono.workspace = true
+proximadb-kernel.workspace = true
+serde.workspace = true
+thiserror.workspace = true
+url = "2.4"
+memmap2 = "0.9"
